Refactor: extract state and serial messages into new files
